+++ pages/views/ingredients/compos.php
@@ -4,7 +4,13 @@ define('__ROOT__', dirname(dirname(dirname(dirname(__FILE__)))));
 
 require_once(__ROOT__.'/inc/sec.php');
 require_once(__ROOT__.'/inc/opendb.php');
+require_once(__ROOT__.'/inc/settings.php');
 
+
+$defPercentage = ($settings['avg'] == $defPercentage) ? "Average percentage used" : 
+         (($settings['min'] == $defPercentage) ? "Minimum percentage used" : 
+         (($settings['max'] == $defPercentage) ? "Maximum percentage used" : ""));
+		 
 $ingID = mysqli_real_escape_string($conn, $_GET["id"]);
 $ingName = mysqli_real_escape_string($conn, $_GET["name"]);
 
@@ -30,7 +36,9 @@ $ingName = mysqli_real_escape_string($conn, $_GET["name"]);
           <th>Name</th>
           <th>CAS</th>
           <th>EINECS</th>
-          <th>Percentage</th>
+          <th>Min percentage</th>
+          <th>Max percentage</th>
+          <th>Average percentage</th>
           <th>IFRA Regulated</th>
           <th>GHS Classification</th>
           <th>Declare</th>
@@ -42,8 +50,8 @@ $ingName = mysqli_real_escape_string($conn, $_GET["name"]);
 
 <script type="text/javascript" language="javascript" >
 $(document).ready(function() {
+	
 
-	$('[data-bs-toggle="tooltip"]').tooltip();
 	var tdCompositions;
  	if ($.fn.DataTable.isDataTable('#tdCompositions')) {
             // Destroy existing DataTable instance
@@ -69,12 +77,17 @@ $(document).ready(function() {
 			  { data : 'name', title: 'Name', render: cmpName },
 			  { data : 'cas', title: 'CAS', render: cmpCAS},
 			  { data : 'ec', title: 'EINECS', render: cmpEC},
-			  { data : 'percentage', title: 'Percentage', render: cmpPerc},
+			  { data : 'min_percentage', title: 'Min percentage' , render: cmpMinPerc},
+			  { data : 'max_percentage', title: 'Max percentage' , render: cmpMaxPerc},
+			  { data : 'avg_percentage', title: 'Avg percentage'},
 			  { data : 'IFRA', title: 'IFRA Regulated%'},
 			  { data : 'GHS', title: 'GHS Classification', render: cmpGHS},
 			  { data : 'toDeclare', title: 'Declare', render: cmpDeclare},
 			  { data : null, title: '', render: cmpActions},		   
 			 ],
+		drawCallback: function ( settings ) {
+			$('[rel=tip]').tooltip();
+		},
 		order: [[ 1, 'asc' ]],
 		lengthMenu: [[20, 50, 100, -1], [20, 50, 100, "All"]],
 		pageLength: 20,
@@ -135,8 +148,12 @@ function cmpEC(data, type, row){
 	return '<i class="ec pv_point_gen" data-name="ec" data-type="text" data-pk="'+row.id+'">'+row.ec+'</i>';    
 };
 
-function cmpPerc(data, type, row){
-	return '<i class="percentage pv_point_gen" data-name="percentage" data-type="text" data-pk="'+row.id+'">'+row.percentage+'</i>';    
+function cmpMinPerc(data, type, row){
+	return '<i class="min_percentage pv_point_gen" data-name="min_percentage" data-type="text" data-pk="'+row.id+'">'+row.min_percentage+'</i>';    
+};
+
+function cmpMaxPerc(data, type, row){
+	return '<i class="max_percentage pv_point_gen" data-name="max_percentage" data-type="text" data-pk="'+row.id+'">'+row.max_percentage+'</i>';    
 };
 
 function cmpGHS(data, type, row){
@@ -182,10 +199,24 @@ $('#tdCompositions').editable({
 
 $('#tdCompositions').editable({
     container: 'body',
-    selector: 'i.percentage',
+    selector: 'i.min_percentage',
     type: 'POST',
     url: "/pages/update_data.php?composition=update&ing=<?=$ingName;?>",
-    title: 'Percentage'
+    title: 'Min percentage',
+	success: function (data) {
+			reload_cmp_data();
+	}
+});
+
+$('#tdCompositions').editable({
+    container: 'body',
+    selector: 'i.max_percentage',
+    type: 'POST',
+    url: "/pages/update_data.php?composition=update&ing=<?=$ingName;?>",
+    title: 'Max percentage',
+	success: function (data) {
+			reload_cmp_data();
+	}
 });
 
 $('#tdCompositions').editable({
@@ -265,14 +296,15 @@ $('#addComposition').on('click', '[id*=cmpAdd]', function () {
 		data: {
 			composition: 'add',
 			allgName: $("#allgName").val(),
-			allgPerc: $("#allgPerc").val(),
+			minPerc: $("#minPerc").val(),
+			maxPerc: $("#maxPerc").val(),
 			allgCAS: $("#allgCAS").val(),
 			allgEC: $("#allgEC").val(),	
 			GHS: $("#GHS").val(),	
 			addToIng: $("#addToIng").is(':checked'),
 			addToDeclare: $("#addToDeclare").is(':checked'),
 			ing: '<?=$ingName?>'
-			},
+		},
 		dataType: 'json',
 		success: function (data) {
 			if (data.success) {
@@ -280,7 +312,8 @@ $('#addComposition').on('click', '[id*=cmpAdd]', function () {
 				$("#allgName").val('');
 				$("#allgCAS").val('');
 				$("#allgEC").val('');
-				$("#allgPerc").val('');
+				$("#minPerc").val('');
+				$("#maxPerc").val('');
 				$("#GHS").val('');
 				reload_cmp_data();
 			}else{
@@ -355,8 +388,12 @@ function reload_cmp_data() {
         	<input class="form-control" name="allgEC" type="text" id="allgEC" />
         </div>
         <div class="mb-3">
-	        <label for="allgPerc" class="form-label">Percentage</label>
-    	    <input class="form-control" name="allgPerc" type="text" id="allgPerc" />
+	        <label for="minPerc" class="form-label">Min percentage</label>
+    	    <input class="form-control" name="minPerc" type="text" id="minPerc" />
+        </div>
+        <div class="mb-3">
+	        <label for="maxPerc" class="form-label">Max percentage</label>
+    	    <input class="form-control" name="maxPerc" type="text" id="maxPerc" />
         </div>
         <div class="mb-3">
 	        <label for="GHS" class="form-label">GHS Classification</label>
@@ -399,8 +436,8 @@ function reload_cmp_data() {
             <input type="file" name="CSVFile" id="CSVFile" class="form-control" />
         </div>
         <hr class="dropdown-divider" />        
-        <p>CSV format: <strong>ingredient,CAS,EINECS,percentage,GHS</strong></p>
-        <p>Example: <em><strong>Citral,5392-40-5,226-394-6,0.15,Skin Irrit. 2-Eye Irrit</strong></em></p>
+        <p>CSV format: <strong>ingredient,CAS,EINECS,min percentage, max percentage,GHS</strong></p>
+        <p>Example: <em><strong>Citral,5392-40-5,226-394-6,0.15,2.3,Skin Irrit. 2-Eye Irrit</strong></em></p>
         <p>Duplicates will be ignored.</p>
             
 	  <div class="modal-footer">