+++ releasenotes.md
@@ -1,5 +1,7 @@
 Whats New in v10.5
 --------------------------
+- Added min and max percentage for compositions
+- Added formula analysis to breakdown sub-materials and percentages
 - Introduse SDS creation (PREVIEW)
 - Added IFRA limit in composition if found in library
 - SDS in ingredient has been renamed to a Document